recovers missing code that got lost in a merge somewhere.  This calls the relevant fixup for infiltration with concentration.  refs #65
