+++ src/pks/mpc/mpc_coupled_transport.cc
@@ -1,14 +1,15 @@
 /*
-  This is the mpc_pk component of the Amanzi code.
+  This is the mpc_pk component of the Amanzi code. 
 
-  Copyright 2010-201x held jointly by LANS/LANL, LBNL, and PNNL.
-  Amanzi is released under the three-clause BSD License.
-  The terms of use and "as is" disclaimer for this license are
+  Copyright 2010-201x held jointly by LANS/LANL, LBNL, and PNNL. 
+  Amanzi is released under the three-clause BSD License. 
+  The terms of use and "as is" disclaimer for this license are 
   provided in the top-level COPYRIGHT file.
 
 */
 
 #include "mpc_coupled_transport.hh"
+
 
 namespace Amanzi {
 
@@ -20,18 +21,20 @@ namespace Amanzi {
     WeakMPC(pk_tree_or_fe_list, global_list, S, soln)
   {
 
-    // Create verbosity object.
+    //std::cout << *global_list<<"\n";
+    //exit(0);
+      // Create verbosity object.
     vo_ = Teuchos::null;
     Teuchos::ParameterList vlist;
     vlist.sublist("verbose object") = global_list -> sublist("verbose object");
-    vo_ =  Teuchos::rcp(new VerboseObject("Coupled TransportPK", vlist));
-
+    vo_ =  Teuchos::rcp(new VerboseObject("Coupled TransportPK", vlist)); 
+    
     name_ = "surface subsurface transport";
 
     Teuchos::Array<std::string> pk_order = plist_->get< Teuchos::Array<std::string> >("PKs order");
-
+  
     for (int i=0; i<2; i++){
-      Teuchos::RCP<Teuchos::ParameterList> list = Teuchos::sublist(Teuchos::sublist(global_list, "PKs"), pk_order[i]);
+      Teuchos::RCP<Teuchos::ParameterList> list = Teuchos::sublist(Teuchos::sublist(global_list, "PKs"), pk_order[i]);      
       Key dom_name = list->get<std::string>("domain name", "domain");
       if (dom_name == "domain"){
         subsurface_transport_list_ = list;
@@ -46,28 +49,25 @@ namespace Amanzi {
       }
     }
 
-    subsurface_flux_key_ =  plist_->get<std::string>("flux_key",
+    subsurface_flux_key_ =  plist_->get<std::string>("flux_key", 
                                                      Keys::getKey(subsurface_name_, "mass_flux"));
 
-    surface_flux_key_ =  plist_->get<std::string>("flux_key",
+    surface_flux_key_ =  plist_->get<std::string>("flux_key", 
                                                   Keys::getKey(surface_name_, "mass_flux"));
 
-
+    
   }
 
 // -----------------------------------------------------------------------------
 // Calculate the min of sub PKs timestep sizes.
 // -----------------------------------------------------------------------------
 double CoupledTransport_PK::get_dt() {
-  //double dt = Amanzi::PK_MPCSubcycled_ATS::get_dt();
-  //ComputeVolumeDarcyFlux(S_next_.ptr());
 
   double surf_dt = sub_pks_[surf_id_]->get_dt();
   double subsurf_dt = sub_pks_[subsurf_id_]->get_dt();
 
   Teuchos::OSTab tab = vo_->getOSTab();
-
-  if (vo_->os_OK(Teuchos::VERB_HIGH)) {
+  if (vo_->getVerbLevel() >= Teuchos::VERB_HIGH){
     *vo_->os()<< "surface transport dt = "<<surf_dt<<"\n";
     *vo_->os()<< "sub surface transport dt = "<<subsurf_dt<<"\n";
   }
@@ -75,7 +75,7 @@ double CoupledTransport_PK::get_dt() {
   double dt = std::min(surf_dt, subsurf_dt);
   set_dt(dt);
   return dt;
-
+ 
 }
 
 
@@ -92,10 +92,14 @@ double CoupledTransport_PK::get_dt() {
 void CoupledTransport_PK::Setup(const Teuchos::Ptr<State>& S){
 
   //passwd_ = "coupled_transport";  // owner's password
-  //passwd_ = "state";  // owner's password
 
   WeakMPC::Setup(S);
 
+  subsurf_pk_ = Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[subsurf_id_]);
+  AMANZI_ASSERT(subsurf_pk_ != Teuchos::null);  
+  surf_pk_ = Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[surf_id_]);
+  AMANZI_ASSERT(surf_pk_ != Teuchos::null);  
+  
 }
 
 void CoupledTransport_PK::Initialize(const Teuchos::Ptr<State>& S){
@@ -103,60 +107,77 @@ void CoupledTransport_PK::Initialize(const Teuchos::Ptr<State>& S){
   WeakMPC::Initialize(S);
 
 }
+int CoupledTransport_PK::num_aqueous_component(){
+
+  int num_aq_comp = subsurf_pk_ -> num_aqueous_component();
+
+  if (num_aq_comp != surf_pk_ -> num_aqueous_component()){
+    Errors::Message message("CoupledTransport_PK:: numbers aqueous component does not match.");
+    Exceptions::amanzi_throw(message);
+  }
+
+  return num_aq_comp;
+
+}
 
 // -----------------------------------------------------------------------------
 // Advance each sub-PK individually, returning a failure as soon as possible.
 // -----------------------------------------------------------------------------
 bool CoupledTransport_PK::AdvanceStep(double t_old, double t_new, bool reinit) {
+
   bool fail = false;
+  double dt_MPC = S_inter_->final_time() - S_inter_->initial_time();
 
-  //double dt_MPC = S_->final_time() - S_->initial_time();
+  AMANZI_ASSERT(subsurf_pk_ != Teuchos::null);  
+  AMANZI_ASSERT(surf_pk_ != Teuchos::null);
+
+  
+  // In the case of rain sources these rain source are mixed with solutes on the surface
+  // to provide BC for subsurface domain
+  surf_pk_ -> MixingSolutesWthSources(t_old, t_new);
+  
+  surf_pk_ -> AdvanceStep(t_old, t_new, reinit);
+  subsurf_pk_ -> AdvanceStep(t_old, t_new, reinit);
 
 
-  sub_pks_[subsurf_id_]->AdvanceStep(t_old, t_new, reinit);
-  *vo_->os() <<"Subsurface step successful\n";
-  sub_pks_[surf_id_]->AdvanceStep(t_old, t_new, reinit);
-  *vo_->os() <<"Overland step successful\n";
-
-
-  const Epetra_MultiVector& surf_tcc = *S_inter_->GetFieldCopyData("surface-total_component_concentration", "subcycling")->ViewComponent("cell",false);
+  const Epetra_MultiVector& surf_tcc = *S_inter_->GetFieldCopyData("surface-total_component_concentration", "subcycling")->ViewComponent("cell",false);  
   const Epetra_MultiVector& tcc = *S_inter_->GetFieldCopyData("total_component_concentration", "subcycling")->ViewComponent("cell",false);
 
-  const std::vector<std::string>&  component_names_sub =
-    Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[subsurf_id_])->component_names();
 
-  int num_components =  Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[subsurf_id_]) -> num_aqueous_component();
-   std::vector<double> mass_subsurface(num_components, 0.), mass_surface(num_components, 0.);
+  const std::vector<std::string>&  component_names_sub = subsurf_pk_ -> component_names();
+   
+  int num_components =  subsurf_pk_ -> num_aqueous_component();
+  
+  std::vector<double> mass_subsurface(num_components, 0.), mass_surface(num_components, 0.);
 
-
-  if (vo_->getVerbLevel() >= Teuchos::VERB_MEDIUM){
+  
+  if (vo_->getVerbLevel() >= Teuchos::VERB_MEDIUM){  
     for (int i=0; i<num_components; i++){
+      *vo_->os() << component_names_sub[i]<< ":";
       mass_subsurface[i] = Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[subsurf_id_])
         ->ComputeSolute(tcc, i);
       mass_surface[i] = Teuchos::rcp_dynamic_cast<Transport::Transport_ATS>(sub_pks_[surf_id_])
         ->ComputeSolute(surf_tcc, i);
       Teuchos::OSTab tab = vo_->getOSTab();
-
-      *vo_->os() <<"subsurface =" << mass_subsurface[i] << " mol";
+      *vo_->os() <<" subsurface =" << mass_subsurface[i] << " mol";
       *vo_->os() <<", surface =" << mass_surface[i]<< " mol";
       *vo_->os() <<", ToTaL =" << mass_surface[i]+mass_subsurface[i]<< " mol" <<std::endl;
     }
   }
-
+  
   return fail;
-
 
 }
 
 
 /* *******************************************************************
-* Interpolate linearly in time between two values v0 and v1. The time
+* Interpolate linearly in time between two values v0 and v1. The time 
 * is measuared relative to value v0; so that v1 is at time dt. The
-* interpolated data are at time dt_int.
+* interpolated data are at time dt_int.            
 ******************************************************************* */
 void CoupledTransport_PK::InterpolateCellVector(
-    const Epetra_MultiVector& v0, const Epetra_MultiVector& v1,
-    double dt_int, double dt, Epetra_MultiVector& v_int)
+    const Epetra_MultiVector& v0, const Epetra_MultiVector& v1, 
+    double dt_int, double dt, Epetra_MultiVector& v_int) 
 {
   double a = dt_int / dt;
   double b = 1.0 - a;
